% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfde_opportunity_numbers.R
\name{cfde_opportunity_numbers}
\alias{cfde_opportunity_numbers}
\title{All CFDE Funding Opportunity Numbers}
\usage{
cfde_opportunity_numbers(
  url = "https://commonfund.nih.gov/dataecosystem/FundingOpportunities"
)
}
\arguments{
\item{url}{The URL of the CFDE funding webpage}
}
\value{
a character vector of funding opportunity numbers.
}
\description{
This function retrieves all CFDE funding
opportunity numbers from the CFDE funding
website, \url{https://commonfund.nih.gov/dataecosystem/FundingOpportunities}.
}
\details{
Note that this function is specific to the CFDE
program and is not a general-purpose web scraping
function.
}
\examples{

\dontrun{
browseURL("https://commonfund.nih.gov/dataecosystem/FundingOpportunities")
}

cfde_opportunity_numbers()

}
