% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_excel}
\alias{export_excel}
\title{Export to Excel}
\usage{
export_excel(
  core_project_numbers,
  token = gitcreds::gitcreds_get()$password,
  service_account_json = "cfde-access-keyfile.json",
  dir
)
}
\arguments{
\item{core_project_numbers}{A character vector of NIH Core Project Numbers}

\item{token}{The token required for authentication with the GitHub API}

\item{service_account_json}{A character string containing the path to a JSON file containing a Google service account}

\item{dir}{A character string containing the path to directory where the Excel file will be written}
}
\description{
Export to Excel
}
\examples{
\dontrun{
test_projects <-c("OT2OD030545")
}

}
