% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_tabular}
\alias{export_tabular}
\title{Export to Tabular}
\usage{
export_tabular(
  core_project_numbers,
  token = gitcreds::gitcreds_get()$password,
  service_account_json = "cfde-access-keyfile.json",
  dir,
  csv = FALSE
)
}
\arguments{
\item{core_project_numbers}{A character vector of NIH Core Project Numbers}

\item{token}{The token required for authentication with the GitHub API}

\item{service_account_json}{A character string containing the path to a JSON file containing a Google service account}

\item{dir}{A character string containing the path to directory where the Excel file will be written}

\item{csv}{A logical indicating whether to write a CSV file}
}
\description{
Export to Tabular
}
\examples{
\dontrun{
test_projects <-c("OT2OD030545")
}

}
