% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_meta_browser.R
\name{ga_meta_simple}
\alias{ga_meta_simple}
\title{Google analytics metadata browser}
\usage{
ga_meta_simple()
}
\description{
This function retrieves Google Analytics metadata
and returns it as a dataframe.
}
\examples{
ga_meta_simple()

}
\seealso{
Other Google Analytics: 
\code{\link{ga_metrics_and_dims_browser}()},
\code{\link{ga_query_explorer}()}
}
\concept{Google Analytics}
