% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_meta_browser.R
\name{ga_metrics_and_dimensions_browser}
\alias{ga_metrics_and_dimensions_browser}
\title{Launch Google Analytics metadata browser Shiny app}
\usage{
ga_metrics_and_dimensions_browser(...)
}
\arguments{
\item{...}{Additional parameters passed to shiny::runApp()}
}
\description{
This function creates a Shiny app that allows users to search and filter
Google Analytics metadata interactively.
}
\examples{
\dontrun{
ga_metrics_and_dimensions_browser()
}

}
\seealso{
Other Google Analytics: 
\code{\link{ga_meta_simple}()},
\code{\link{ga_query_explorer}()}
}
\concept{Google Analytics}
