% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_by_topic.R
\name{get_contributor_count}
\alias{get_contributor_count}
\title{Get the number of contributors for a GitHub repository}
\usage{
get_contributor_count(owner, repo, token = NULL)
}
\arguments{
\item{owner}{The owner of the GitHub repository}

\item{repo}{The name of the GitHub repository}

\item{token}{A GitHub personal access token (optional)}
}
\value{
The number of contributors to the repository
}
\description{
Get the number of contributors for a GitHub repository
}
