% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih_reporter.R
\name{get_core_project_info}
\alias{get_core_project_info}
\title{Get Core Project Info}
\usage{
get_core_project_info(core_project_numbers)
}
\arguments{
\item{core_project_numbers}{A character vector of core project numbers.}
}
\value{
A tibble with the following columns:
\describe{
\item{core_project_number}{The core project number.}
\item{found_publication}{A logical indicating whether any publications were found for this core project number.}
\item{appl_id}{The application ID associated with the publication.}
\item{pmid}{The PubMed ID associated with the publication.}
\item{pubmed_url}{The URL of the PubMed page for the publication.}
}
}
\description{
Given NIH Core Project Number(s), this function retrieves project metadata
including any assosciated publications from the NIH RePORTER API.
}
\examples{
# Get publications for a set of core project numbers
proj_info <- get_core_project_info(c("u54od036472", "99999999"))
# View the results
proj_info
}
