% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ga_basic.R
\name{get_ga_basic}
\alias{get_ga_basic}
\title{Get Basic Google Analytics Info}
\usage{
get_ga_basic(
  core_project_numbers,
  service_account_json = "cfde-access-keyfile.json"
)
}
\arguments{
\item{core_project_numbers}{A character vector of NIH Core Project Numbers}

\item{service_account_json}{A character string containing the path to a JSON file containing the
Google service account credentials. Defaults to "cfde-access-keyfile.json"}
}
\value{
A data frame containing the associated Google Analytics data
}
\description{
This function takes a character vector of NIH Core Project Numbers and
returns a data frame containing the any Google Analytics properties associated
with the Core Project Numbers.
}
