% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_meta_simple.R
\name{get_ga_meta_by_id}
\alias{get_ga_meta_by_id}
\title{Google Analytics metadata dataframe by property ID}
\usage{
get_ga_meta_by_id(property_id)
}
\arguments{
\item{property_id}{The property ID for which to retrieve
metadata.}
}
\value{
A tibble containing Google Analytics metadata.
}
\description{
This function retrieves Google Analytics metadata by property ID
and returns it as a dataframe. The metadata includes
information about metrics, dimensions, and other
attributes available in Google Analytics.
}
\details{
This function is a wrapper around the
\code{googleAnalyticsR::ga_meta()} function. It retrieves
metadata for the Google Analytics API version 4.
}
\note{
This function requires first authenticating to
Google Analytics using the \code{ga_auth()} function.
}
\examples{
\dontrun{
res = get_ga_meta_by_id("123456789")
head(res)
dplyr::glimpse(res)
}

}
\seealso{
Other Google Analytics: 
\code{\link{ga_meta_simple}()},
\code{\link{ga_query_explorer}()}
}
\concept{Google Analytics}
