% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_by_topic_graphQL.R
\name{get_github_by_topic_graphql}
\alias{get_github_by_topic_graphql}
\title{Get GitHub Repositories by Topic (GraphQL API)}
\usage{
get_github_by_topic_graphql(topics, token, limit = 30)
}
\arguments{
\item{topics}{A vector of GitHub topics to search for.}

\item{token}{A GitHub personal access token. Required for GraphQL API.}

\item{limit}{The maximum number of repositories to return per topic (max 1000 by GitHub).}
}
\value{
A data frame with repository metadata.
}
\description{
Get GitHub Repositories by Topic (GraphQL API)
}
\examples{
\dontrun{
df <- get_github_by_topic_graphql(c("u24ca289073"), token = "ghp_yourPAT")
dplyr::glimpse(df)
}

}
