% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_topics.R
\name{get_github_topics}
\alias{get_github_topics}
\title{Get Repositories from GitHub by Topic}
\usage{
get_github_topics(topics, token = NULL, limit = 30)
}
\arguments{
\item{topics}{A vector of GitHub topics to search for.}

\item{token}{A GitHub personal access token. If not provided, the function
will run without authentication.}

\item{limit}{The maximum number of results to return. Defaults to 30.}
}
\value{
A data frame containing the results of the search query.
}
\description{
Get Repositories from GitHub by Topic
}
