% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nih_reporter.R
\name{get_publications_for_core_projects}
\alias{get_publications_for_core_projects}
\title{Get Publications for Core Projects}
\usage{
get_publications_for_core_projects(core_project_numbers)
}
\arguments{
\item{core_project_numbers}{A character vector of core project numbers.}
}
\value{
A tibble with the following columns:
\describe{
\item{core_project_number}{The core project number.}
\item{found}{A logical indicating whether any publications were found for this core project number.}
\item{applid}{The application ID associated with the publication.}
\item{pmid}{The PubMed ID associated with the publication.}
\item{pubmed_url}{The URL of the PubMed page for the publication.}
}
}
\description{
This function retrieves a list of publications associated with a set of core
project numbers from the NIH RePORTER API.
}
