% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_excel}
\alias{export_excel}
\title{Export to Excel}
\usage{
export_excel(core_project_numbers, token = gitcreds::gitcreds_get()$password)
}
\description{
Export to Excel
}
\examples{
\dontrun{
test_projects <-c("OT2OD030545")
}

}
