% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_meta_browser.R
\name{ga_query_explorer}
\alias{ga_query_explorer}
\title{Browse google analytics query explorer}
\usage{
ga_query_explorer()
}
\value{
Opens the Google Analytics Query Explorer in a web browser.
}
\description{
This function opens the Google Analytics Query Explorer in a web browser.
}
\details{
It allows users to explore and test Google Analytics queries interactively.
This is useful for developers and analysts who want to experiment with
different queries and see the results in real-time.
}
\note{
This function is intended for users who have a Google Analytics account
and want to explore the API capabilities. It does not return any data or
perform any analysis.
}
\examples{
\dontrun{
ga_query_explorer()
}

}
\seealso{
Other Google Analytics: 
\code{\link{ga_meta_simple}()}
}
\concept{Google Analytics}
